requirements.txt
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.2
beniget==0.4.1
bleach==6.0.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
cliapp==1.20180812.1
click==8.0.3
colorama==0.4.4
comm==0.1.2
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
duplicity==0.8.21
email-validator==1.3.1
entrypoints==0.4
executing==1.2.0
fastapi==0.92.0
fasteners==0.14.1
fastjsonschema==2.16.3
fonttools==4.38.0
fqdn==1.5.1
fs==2.4.12
future==0.18.2
gast==0.5.2
graphviz==0.20.1
greenlet==1.1.2
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httplib2==0.20.2
httptools==0.5.0
httpx==0.23.3
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.2
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.20.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
multimethod==1.9.1
mysql-connector-python==8.0.33
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.0
olefile==0.46
orjson==3.8.7
packaging==23.0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.1.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydantic==1.10.5
pydot==1.4.2
pydotplus==2.0.2
pygame==2.3.0
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
pythomata==0.3.2
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
reportlab==3.6.8
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.1
scikit-learn==1.2.2
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
sly==0.5
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.25.0
sympy==1.9
systemd-python==234
tabulate==0.9.0
terminado==0.17.1
testpath==0.5.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
ttystatus==0.38
typer==0.7.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.7.0
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.2.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.20.0
uvloop==0.17.0
wadllib==1.3.6
watchfiles==0.18.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
widgetsnbextension==4.0.5
xdg==5
xkit==0.0.0
y-py==0.5.9
ypy-websocket==0.8.2
zipp==1.0.0
